+++ FirstRealApp/Controllers/ImagesController.cs
@@ -1,6 +1,6 @@
-﻿using FirstRealApp.Interfaces;
-using FirstRealApp.Models;
+﻿using FirstRealApp.Models;
 using FirstRealApp.Repository;
+using FirstRealApp.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
