+++ FirstRealApp/Services/ServiceResponse.cs
@@ -1,0 +1,12 @@
+﻿namespace FirstRealApp.Services
+{
+    public class ServiceResponse<T>
+    {
+        public T? Data { get; set; }
+
+        public bool Success { get; set; }
+
+        public string? Message { get; set; }
+
+    }
+}
